+++ controllers/llanta.ts
@@ -2,12 +2,14 @@ import {Request, Response} from 'express';
 import { deleteImage, uploadImage } from '../libs/cloudinary';
 import fs from 'fs-extra';
 import { LLanta } from '../models';
+import { getFecha } from '../helpers';
 
 
 export const getLlantas = async (req: Request, res: Response)=>{
 
     try {
         const productos = await LLanta.findAll();
+        console.log("productos", productos)
         return res.json({
             productos
         });
@@ -44,10 +46,11 @@ export const getLlanta = async (req: Request, res: Response)=>{
 
 export const createLlanta = async (req: Request, res: Response)=>{
 
-    console.log("req.body desde el controlador", req.body);
         try {
 
-        const {id_producto, cantidad, marca, precio, stock, descripcion} = req.body;
+        const {id_producto, marca, precio, stock, descripcion} = req.body;
+        const { fecha, times_created } = getFecha();
+
         let image;
         let image_public_id;
 
@@ -64,7 +67,17 @@ export const createLlanta = async (req: Request, res: Response)=>{
                 });
             }
 
-            if(req.files!.imagen){
+            if(req.files === null){
+                await LLanta.create({
+                    id_producto,
+                    marca: marca.split('')[0].toUpperCase() + marca.slice(1),
+                    precio: parseFloat(precio),
+                    stock,
+                    descripcion,
+                    created_at: fecha,
+                    times_created,
+                })
+            } else {
                 const result = await uploadImage(req.files!.imagen.tempFilePath);
                
                 await fs.remove(req.files!.imagen.tempFilePath);
@@ -72,18 +85,19 @@ export const createLlanta = async (req: Request, res: Response)=>{
                 image = result.secure_url;
     
                 image_public_id = result.public_id;
-            }
 
-            await LLanta.create({
-                id_producto,
-                marca: marca.split('')[0].toUpperCase() + marca.slice(1),
-                cantidad,
-                precio: parseFloat(precio),
-                stock,
-                descripcion,
-                imagen: image,
-                imagen_public_id: image_public_id
-            })
+                await LLanta.create({
+                    id_producto,
+                    marca: marca.split('')[0].toUpperCase() + marca.slice(1),
+                    precio: parseFloat(precio),
+                    stock,
+                    descripcion,
+                    imagen: image,
+                    imagen_public_id: image_public_id,
+                    created_at: fecha,
+                    times_created,
+                })
+            }
 
             res.json({
                 msg: `Producto registrado exitosamente!`
@@ -119,31 +133,46 @@ export const updateLlanta = async (req: Request, res: Response)=>{
             });
         }
 
-        await deleteImage(producto.dataValues.imagen_public_id)
+        if(req.files === null){
+            await producto.update( 
+                {
+                    marca: marca.split('')[0].toUpperCase() + marca.slice(1),
+                    precio: parseFloat(precio),
+                    stock,
+                    descripcion,
+                }, 
+                { 
+                    where: {
+                        id: id_producto,
+                    }
+                }
+            );
+        } else {
+            if(producto.dataValues.imagen_public_id){
+                await deleteImage(producto.dataValues.imagen_public_id)
+            }   
 
-        if(req.files!.imagen){
             const result = await uploadImage(req.files!.imagen.tempFilePath);
             await fs.remove(req.files!.imagen.tempFilePath);
             image = result.secure_url;
             image_public_id = result.public_id;
-        }
 
-        await producto.update( 
-            {
-                marca: marca.split('')[0].toUpperCase() + marca.slice(1),
-                precio: parseFloat(precio),
-                stock,
-                descripcion,
-                imagen: image,
-                imagen_public_id: image_public_id,
-            }, 
-            { 
-                where: {
-                    id: id_producto,
+            await producto.update( 
+                {
+                    marca: marca.split('')[0].toUpperCase() + marca.slice(1),
+                    precio: parseFloat(precio),
+                    stock,
+                    descripcion,
+                    imagen: image,
+                    imagen_public_id: image_public_id,
+                }, 
+                { 
+                    where: {
+                        id: id_producto,
+                    }
                 }
-            }
-        );
-        
+            );
+        }
 
         res.json( {
             msg: "Producto actualizado correctamente",
